+++ service/src/main/java/org/ehrbase/service/TenantServiceImp.java
@@ -17,6 +17,7 @@
  */
 package org.ehrbase.service;
 
+import static org.ehrbase.api.tenant.TenantAuthentication.DEFAULT_TENANT_ID;
 import static org.ehrbase.dao.access.jooq.TenantAccess.getSysTenants;
 
 import java.util.List;
@@ -25,6 +26,7 @@ import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 import org.apache.commons.lang3.StringUtils;
 import org.ehrbase.api.definitions.ServerConfig;
+import org.ehrbase.api.exception.ObjectNotFoundException;
 import org.ehrbase.api.service.TenantService;
 import org.ehrbase.api.tenant.Tenant;
 import org.ehrbase.api.tenant.TenantAuthentication;
@@ -37,6 +39,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.cache.Cache;
 import org.springframework.cache.CacheManager;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -54,6 +57,7 @@ public class TenantServiceImp extends BaseServiceImp implements TenantService {
     private static final String ERR_GETTING_TENANT = "Could not find tenant %s in cache";
     private static final String WARN_NOT_TENANT_ID =
             "No tenant identifier provided, falling back to default tenant identifier {}";
+    private static final String ERR_DELETE_TENANT_ID = "Deleting tenant id[%s] is not allowed";
 
     private final Cache sysTenantCache;
 
@@ -113,6 +117,22 @@ public class TenantServiceImp extends BaseServiceImp implements TenantService {
         return tenant.getTenantId();
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
+    @Override
+    public void deleteTenant(String tenantId) {
+        if (DEFAULT_TENANT_ID.equals(tenantId)) {
+            throw new IllegalArgumentException(
+                    String.format("Not allowed to delete default tenant with id %s.", tenantId));
+        }
+
+        if (!hasTenant(tenantId)) {
+            throw new ObjectNotFoundException("Tenant", String.format("Tenant with id %s does not exist.", tenantId));
+        }
+
+        I_TenantAccess.deleteTenant(getDataAccess().getContext(), tenantId);
+        sysTenantCache.evict(tenantId);
+    }
+
     @Override
     public List<Tenant> getAll() {
         return I_TenantAccess.getAll(getDataAccess().getContext()).stream()
@@ -131,5 +151,10 @@ public class TenantServiceImp extends BaseServiceImp implements TenantService {
     public Tenant update(Tenant tenant) {
         return I_TenantAccess.retrieveInstanceBy(getDataAccess().getContext(), tenant.getTenantId())
                 .update(tenant);
+    }
+
+    @Override
+    public boolean hasTenant(String tenantId) {
+        return I_TenantAccess.hasTenant(getDataAccess(), tenantId);
     }
 }